Reimplement math abstractions

+++ src/main/java/me/deftware/client/framework/entity/types/LivingEntity.java
@@ -31,7 +31,7 @@ public class LivingEntity extends Entity {
 		return getMinecraftEntity().getMaxHealth();
 	}
 
-	public void setMovementMultiplier(float multiplier) {
+	public void setMovementMultiplier(double multiplier) {
 		// getMinecraftEntity().airStrafingSpeed = multiplier; TODO: FIXME
 	}
